refactor(Lotto): model 패키지로 이동
