+++ src/main/java/lotto/model/WinningNumbers.java
@@ -2,7 +2,6 @@ package lotto.model;
 
 import java.util.ArrayList;
 import java.util.List;
-import lotto.Lotto;
 import lotto.constant.Prize;
 
 public class WinningNumbers {
